refactor: update va-progress-bar and va-progress-circle paths.
