+++ packages/ui/src/components/va-button/VaButton.vue
@@ -59,7 +59,7 @@ import { useRouterLink, useRouterLinkProps } from '../../composables/useRouterLi
 import { useSizeProps, useSize } from '../../composables/useSize'
 import { useLoadingProps } from '../../composables/useLoading'
 import VaIcon from '../va-icon'
-import { VaProgressCircle } from '../va-progress-bar'
+import { VaProgressCircle } from '../va-progress-circle'
 
 export default defineComponent({
   name: 'VaButton',
