+++ packages/ui/src/vuestic-plugin/vuestic-components.ts
@@ -49,10 +49,8 @@ export { default as VaNavbar, VaNavbarItem } from '../components/va-navbar'
 export { default as VaPagination } from '../components/va-pagination'
 export { default as VaParallax } from '../components/va-parallax'
 export { default as VaPopover } from '../components/va-popover'
-export {
-  VaProgressBar,
-  VaProgressCircle,
-} from '../components/va-progress-bar'
+export { VaProgressBar } from '../components/va-progress-bar'
+export { VaProgressCircle } from '../components/va-progress-circle'
 export { default as VaRadio } from '../components/va-radio'
 export { default as VaRating } from '../components/va-rating'
 export { default as VaScrollbar } from '../components/va-scrollbar'
